requirements.txt
numpy==1.16.4
Pillow==6.1.0
dataset==1.1.2
facenet_pytorch==0.1.0
Flask==1.1.1
flask_restplus==0.13.0
joblib==0.13.2
matplotlib==3.0.0
seaborn==0.9.0
scikit_learn==0.21.3
Werkzeug==0.15.2
opencv-python==4.1.0.25
uWSGI==2.0.18
